Creating a neat folder for algorithms, each algorithm has its own file, there is a file that is the abstract class of the algorithms and there is a file that creates the algorithm by name.
